+++ config/config.exs
@@ -20,6 +20,8 @@ config :elixir_boilerplate, ElixirBoilerplate.Gettext, default_locale: "en"
 
 config :elixir_boilerplate, ElixirBoilerplateWeb.Plus.Security, allow_unsafe_scripts: false
 
+config :elixir_boilerplate, ElixirBoilerplateGraphQL, enable_introspection: false
+
 config :esbuild,
   version: "0.14.41",
   default: [