+++ lib/elixir_boilerplate_graphql/schema.ex
@@ -3,6 +3,8 @@ defmodule ElixirBoilerplateGraphQL.Schema do
 
   alias ElixirBoilerplate.Repo
 
+  @pipeline_modifier ElixirBoilerplateGraphQL.Pipeline.Introspection
+
   import_types(Absinthe.Type.Custom)
   import_types(ElixirBoilerplateGraphQL.Application.Types)